# docs/openapi.yaml
openapi: 3.1.0
info:
  title: TechXayan Creative Video Editor API
  version: 1.0.0
  description: |
    Production-grade enterprise backend API for TechXayan Creative / my_editor (Windows + Android).
    Provides authentication, project sync with optimistic concurrency control, media asset ingestion
    and intelligence (vector/semantic search), provider-agnostic AI gateway, asynchronous BullMQ job processing,
    speech-to-text with word-level timing, natural-language editor command orchestration,
    realtime collaboration & review timecode commenting, WebSocket/SSE domain event streaming,
    HMAC-verified webhook ingestion with dead-letter queue, and Prometheus observability metrics.
servers:
  - url: http://localhost:4000
    description: Local Development / Test Server
  - url: https://api.editor.techxayan.com
    description: Production Cluster
paths:
...
  /v1/media/register:
    post:
      summary: Register pre-existing / local media asset
      description: Register a media asset record with client-provided technical telemetry, skipping remote transcoding if already suitable.
      operationId: registerMedia
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [fileName, mimeType, fileSizeBytes]
              properties:
                fileName: { type: string }
                mimeType: { type: string }
                fileSizeBytes: { type: integer }
                category: { type: string, enum: [video, audio, image, font, lut, sticker, template] }
                folderId: { type: string }
                checksumSha256: { type: string, description: "64 hex SHA-256 hash" }
                durationSeconds: { type: number }
                width: { type: integer }
                height: { type: integer }
                framerate: { type: number }
                codec: { type: string }
                bitrateKbps: { type: integer }
                audioChannels: { type: integer }
                audioSampleRate: { type: integer }
                rotation: { type: integer }
      responses:
        '201':
          description: Media asset registered with READY status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaAsset'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    AdminAuth:
      type: apiKey
      in: header
      name: x-admin-key
  schemas:
    MediaAsset:
      type: object
      properties:
        id: { type: string }
        userId: { type: string }
        projectId: { type: string, nullable: true }
        folderId: { type: string, nullable: true }
        name: { type: string }
        originalFilename: { type: string }
        category: { type: string, enum: [video, audio, image, font, lut, sticker, template] }
        fileKey: { type: string }
        mimeType: { type: string }
        fileSizeBytes: { type: integer }
        checksumSha256: { type: string }
        durationSeconds: { type: number }
        width: { type: integer }
        height: { type: integer }
        framerate: { type: number }
        codec: { type: string }
        bitrateKbps: { type: integer }
        audioChannels: { type: integer }
        audioSampleRate: { type: integer }
        rotation: { type: integer }
        orientation: { type: string, enum: [landscape, portrait, square] }
        status: { type: string, enum: [UPLOADING, UPLOADED, PROCESSING, READY, FAILED, DELETING, DELETED, ARCHIVED] }
        isFavorite: { type: boolean }
        downloadUrl: { type: string, format: uri }
        thumbnailUrl: { type: string, format: uri }
        variants:
          $ref: '#/components/schemas/MediaVariants'
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    MediaVariants:
      type: object
      properties:
        original:
          type: object
          properties:
            fileKey: { type: string }
            url: { type: string }
            sizeBytes: { type: integer }
            mimeType: { type: string }
        thumbnail:
          type: object
          properties:
            fileKey: { type: string }
            url: { type: string }
            width: { type: integer }
            height: { type: integer }
        waveform:
          type: object
          properties:
            fileKey: { type: string }
            url: { type: string }
            peaks:
              type: array
              items: { type: number }
            channels: { type: integer }
            samplesPerPixel: { type: integer }
        proxy:
          type: object
          properties:
            fileKey: { type: string }
            url: { type: string }
            resolution: { type: string }
            codec: { type: string }
            sizeBytes: { type: integer }
        previewDerivative:
          type: object
          properties:
            fileKey: { type: string }
            url: { type: string }
            resolution: { type: string }
            durationSeconds: { type: number }
...